requirements.txt
googlemaps==2.5.1
weather-api==1.0.4
webbrowser
